Full template for the home page.
There is supposed to be .IsHome node, but it doesn't work. To work around it, I
need to create a full <html>…</html> template for the home page.

A good thing about it is that now only the home page contains the RSS link,
instead of each page containing that.