Add full article body to JSON-LD.
I'm slightly confused about this. The schema.org article about the
Article object says that the type here is "Text". All right, but what
does it mean? Only plaintext allowed?

For now, I've stripped all the HTML tags and escaped what was left, to
take care of all the quotation marks, etc. For clean escaping, it seems
hard to mix HTML embedded in Markdown with the HTML generated by the
Markdown to HTML transformation.