Use JavaScript for the Performances page.
Using the client's local date instead of the server generation date.

This will work incorrectly when used in other time zones – the script
assumes the local time zone. Or, in the absence of other information,
JavaScript, or the Date library, assumes the local time zone. I might
fix it some time, but for these purposes it works well enough. I just
want events not to linger as "future" when it's days or weeks past the
date.